fix: make min width limit work for other screen sizes
JIRA: LX-668
risk: low
